requirements.txt
flask==3.1.0
flask-login==0.6.3
flask-sqlalchemy==3.1.1
gunicorn==23.0.0
numpy==2.2.4
pandas==2.2.3
psycopg2-binary==2.9.10
scikit-learn==1.6.1
werkzeug==3.1.3
sqlalchemy==2.0.40
openai==1.75.0
trafilatura==2.0.0
slack-sdk==3.35.0
twilio==9.5.2
email-validator==2.2.0
python-dotenv==1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
sentence-transformers>=2.2.2
duckduckgo-search>=4.1.1
ollama>=0.1.0
requests>=2.31.0